rpc: move ContractDetails to request/ package
